fix: published version in apikey
